封装参数查询辅助
