+++ README.md
@@ -71,6 +71,9 @@ V 1.4 更新日志:
 Software using Demo:	
 ===
 ```
+python WeblogicScan.py -u 127.0.0.1 -p 7001
+```
+```
 __        __   _     _             _        ____
 \ \      / /__| |__ | | ___   __ _(_) ___  / ___|  ___ __ _ _ __
  \ \ /\ / / _ \ '_ \| |/ _ \ / _` | |/ __| \___ \ / __/ _` | '_ \
